Pulled word lists out into global namespace.
With the last change to main() (allowing for more than one phrase to be
generated in a single run of the script) I figured it was time to pull the
adjective and noun lists out of the function and into global namespace. It was
nice while it lasted. Nice and readable. But when you define big objects inline,
one day you're gonna be sad. When you have to refactor those objects out in
order to not regenerate them on each call. I'll miss you, inline object
definitions.

*half-heartedly kicks nearby paint can*